deployment/requirements.txt
streamlit
scikit-learn == 1.1.2
tensorflow-cpu
pandas
numpy
